Make autofix sort keys in objects automatically, fix support for arrays
